First stab at libudev-compat's new monitoring system. The key difference between libudev and libudev-compat is that for "udev" monitors, libudev-compat creates a directory into which the device manager is expected to atomically put (i.e. through hard-link) files that contain data that can be parsed by udev_device_new_from_nulstr() into a struct udev_event instances.  Instead of using a netlink multicast group, a struct udev_monitor sets up an epoll handle that watches the readiness of both a socket (from a socketpair) and an inotify handle set to emit a one-shot event when a new file is added to the device event directory. When the epoll handle is ready (which the client determines by poll()'ing on it, since it is the handle returned by udev_monitor_get_fd()), subsequent calls to udev_monitor_receive_device() reset the inotify handle on the directory, and consume all files in it.  Each file is parsed into a struct udev_device, and fed into a socket from a socketpair (the other end of which is attached to the epoll handle).  This way, the epoll handle polls ready for reading when there is either a new file in the directory (i.e  at least one new event pending), or parsed devices bufferred in the socket, or both.  We implement device filtering by attaching a BPF program to the socket, similar to as before.
Handling the case when the libudev client fork()s is challenging, but
feasible.  We maintain a global table of all monitors in existence,
which we maintain in calls to udev_monitor_new_from_netlink() and
udev_monitor_unref().  Upon fork(), we make a pass over this global table
of struct udev_monitors and atomically re-target all inotify handles to a new
events directory (named to include the PID of the child process).  We
additionally regenerate each monitor's socketpair, and re-attach the
filter, thereby reconstructing the epoll handle.  This way, both the
parent and child process will receive subsequent device events,
emulating the netlink multicast group as before in a way that is
consistent with (but not necessarily identical to) a possible program
execution order with vanilla libudev.

The device manager, be it vdev, udev, mdev, etc., will need to generate
files at the appropriate places.  However, this is trivial in all three
of these device managers, since the logic to do so can be encoded in a
device rule.